Refactor Horde_Db injectors.
Forgot that individual subsystems could potentially define a custom SQL
configuration.  Therefore, we need a Horde_Db injector with a
factory/singleton to override the base DB for these subsystems (and
reuse the base DB object if no custom config is given).

Rename Horde_Db_Adapter_Base to Horde_Db_Base and change to a shortcut
to return the base Horde DB object.