Hotfix for issue that prevents form submissions due to sessions being cleared.
If submitting a form as guest, the secret tokens and captchas etc... are cleared from the session
during the check for transparent authentication.  Assume that if this form is submitted, the user
was already authenticated (or did not require authentication).

As mentioned, this is a hotfix meant to get bugs.horde.org working again for guests.
There is probably a more appropriate way to fix this, but I cannot think of any other fixes that do not
open up other security holes.