Throw error if config file is not found.
If config file is not found at all, this should be treated the same as
errors in the config file - by throwing an Exception (the existing
comment agrees with this).  The calling code should be responsible for
determining whether the absence of the file should be fatal or not (If
fatal, the missing file error message will be logged there).