Fix Horde_Cache_Session creation.
Not sure how else to do this, but with the recent compress config option
added to horde/Cache, this is the only way to pickup this option. Also,
this was broken since theoretically use_memorycache could have been set
to true but creating Horde_Cache_Session this way would not have honored
that config setting.