Don't cache combined config arrays.
This was leading to stale Horde config data being in an app's cached config,
causing incorrect URLs to be generated. If this isn't the correct solution, then
I think application config cache keys need to take an md5sum of the Horde config
into account so they are thrown out when they don't match.