Optimize session storage.
Remove separate serialized index array, and instead track whether entry
is serialized directly within an entry key.

This means we may have to look up two separate keys to find a value, but
we needed to do a lookup anyway previously so this isn't any less
efficient.

Especially with longer key names, this saves a decent amount of space.
I saw a 16% drop in session size after implementing this.