Add Horde_Core_Alarm_Handler_Notify class
The injector passed to the Alarm Notify class requires that create()
have no arguments.  However, the factory class requires that create()
have a Horde_Injector argument.

The alternative is to create a wrapper object that wraps the create()
method and adds the argument.  But at that point, we might as well
create a local Horde_Core driver.

Wondering if we should do this for the Mail and Desktop drivers also.