Improve performance by not pinging for flag changes.
The most important change in a PING request would be for a newly
arrived message in the mailbox. Flag changes are not included in
the PING check to avoid PING/SYNC/PING cycles - especially while
reading mail on the device. Flag changes are still included in SYNC
response, but only after a new mail arrives to trigger the SYNC request.