Bug #11103: Re-add delay to showing notifications on page load
This event is being fired too quickly, so the history is not being
updated before switching to the notification view. This was completely
breaking the ability to use the smartmobile view on actual smartmobile
devices (e.g. Galaxy Nexus), since closing the notification window was
returning to the previous page in the history - the login page - instead
of the previous view in the current page.

This is a total hack, but I couldn't figure out any other way of doing
this in either jquery or jquery mobile.  Not sure if this has been fixed
for 1.1.