Use a specialized factory for ActiveSync loggers.
Since in EAS 12.1, the device id can be sent as part of the base64 enoded binary
data, we won't know it until after the request handling starts. Use a
factory for the logger so we can ensure we log all data to device specific files
if desired. Still not decided if I want device specific folders, with per-request
files or not for debugging.