Fix API method determination during initialization
Looks to be a limitation with PHP's __get() magic method.  If you call a
function inside of __get() that, in turn, calls the same __get() method
name again, this is flagged as an unsuccessful parameter lookup and this
second lookup incorrectly fails.

Workaround by defining method calls to return disabled, link, method,
and noPerms information.