Don't attempt to get the entire folder list if all we need is the type.
When fetching a message from a collection, we only need to know if
a collection type is a calendar, addressbook, email etc... we
don't need to know the exact type of email folder (sent, inbox, etc...).
This potentially saves a LIST call to the IMAP server for each email message
sent to the device.