Use callByPackage().
If an application's links method contains class names from that application,
they will not be able to be autoloaded if the application is not pushed on
the stack. Not sure if using callByPackage() is the best/right fix for this,
but consider this a hotfix so I can access IMP's preferences now.