Fix 'activesync_replyposition' pref readout
The moment we read out the 'activesync_replyposition'
in the constructor was before the user is authenticated.

Therefore the prefs backend will return the default value,
which is 'bottom'. Setting the pref to 'top' via the UI never worked.

Fix it by delaying the read out until the reply/forward action.
Keep "reply_top" parameter handling as compatible as possible.

Bug 12636

Signed-off-by: Michael J Rubinsky <mrubinsk@horde.org>