More atomic control over if a collection is multiplexed or not.
The ability to use multiple collections is going to be pretty useless
unless we can turn it on/off per collection since different clients
have different levels of support for this.

Really want this to be user-configurable per-device but do to a
chicken/egg problem this won't work. I will be adding, instead some
rudimentry device capability sniffing to at least force mulitplexing
on devices that absolutely don't support it.