Add device for Horde_ActiveSync_Message_Contact
Horde_ActiveSync_Message_Contact is checking if _device is available. If it is, the device-specific date format should be used.
This was code was never called, as the _device variable is never set. This should fix it.

Now the device specific date format is used also, which fixes the birthdates of contacts now. Before the dates were shifted by one day.
When editing the birthdate of a contact in the addressbook (for example 04.03.1991), before the log showed: <POOMCONTACTS:Birthday>1991-03-03T23:00:00.000Z</POOMCONTACTS:Birthday>, now it shows the correct device specific date format: <POOMCONTACTS:Birthday>1991-03-04T12:00:00.000Z</POOMCONTACTS:Birthday>