[mms] Horde_Data now stores temporary data using Horde_Vfs to ensure data persists and that it doesn't overload session storage (Request #11335)
Horde_Core_HashTable_PersistentSession has been added to provide a
solution to allow for "persistent" session storage. This means that the
data is guaranteed to exist for the life of the session (but no more).

This fixes the potential of destroying a session by importing too much
data (e.g. in Turba Import screen).