fix selection of mode on login
When the User selects a mode on the login form (ie. setting horde_select_view),
it gets always overwritten by 'auto'. The reason is that during login there is
a cycle that leads to authentication being triggered twice.

The culprit is the call to setAuth at
https://github.com/horde/Core/blob/master/lib/Horde/Core/Auth/Application.php#L669
which ends up triggering a second authentication through

    $this->loadPrefs($this->getApp());

which in turn does not have access to the original $credentials array and then
sets the view to the default value.

The fix is to set the view first.